♻️ refactor(topic): share metadata update schema
