+++ packages/types/src/device.ts
@@ -1,3 +1,5 @@
+import { z } from 'zod';
+
 export type ProjectSkillScope = 'device' | 'project';
 export type ProjectSkillSource = '.agents/skills' | '.claude/skills';
 
@@ -634,3 +636,36 @@ export interface DeviceListProjectSkillsResult {
   /** Legacy source hint. Per-skill `scope` / `source` fields are authoritative. */
   source: ProjectSkillSource | null;
 }
+
+const gitLinkedPullRequestSchema = z.object({
+  ciStatus: z.enum(['failure', 'pending', 'success', 'unknown']).optional(),
+  isDraft: z.boolean().optional(),
+  mergeable: z.string().optional(),
+  mergeStateStatus: z.string().optional(),
+  mergedAt: z.string().nullable().optional(),
+  number: z.number(),
+  reviewDecision: z.string().optional(),
+  state: z.string(),
+  title: z.string(),
+  url: z.string(),
+});
+
+export const workingDirConfigSchema = z.object({
+  git: z
+    .object({
+      activeWorktree: z.string().optional(),
+      branch: z.string().optional(),
+      detached: z.boolean().optional(),
+      github: z
+        .object({
+          extraPullRequestCount: z.number().optional(),
+          pullRequest: gitLinkedPullRequestSchema.nullable().optional(),
+          pullRequestStatus: z.enum(['error', 'gh-missing', 'ok']).optional(),
+        })
+        .optional(),
+      isWorktree: z.boolean().optional(),
+    })
+    .optional(),
+  path: z.string(),
+  repoType: z.enum(['git', 'github']).optional(),
+});
